Remove Crosvm Nvidia workaround
Nvidia's EGL library will fork to run nvidia-modprobe on the
first invocation on a fresh GCE instance. Crosvm interprets
the forked modprobe child process completing as the GPU process
failing which causes failures on the first launch of a fresh
GCE instance. A workaround was added in place to launch the
nvidia-modprobe command on all launch_cvd invocations. We can
now remove this because the cuttlefish-common pacakge installed
on newer host images will run the nvidia-modprobe command on
init.

Bug: b/150633183
Test: launch_cvd --gpu_mode=gfxstream (on fresh GCE instance)
Change-Id: I6e1f7e80e291d4ab95e6148ee5d2837692ba1fac