Track checksums of bin/ and lib64/ files.
This allows the cuttlefish host code to determine whether the existing
cuttlefish_config.json file was generated by the same host binaries or a
different version. More specifically, it can choose based on this value
whether to reuse or replace runtime-generated files, unless the user
explicitly requests one or the other behaviors.

As of this change, only a log message is affected by the host tools
version changing.

Bug: 159068082
Test: launch_cvd multiple times
Change-Id: Ie354d846b343bb016d9ba43b7b1cf71d2d25cd1c