VmManager refactoring
Make VmManager subclasses stateless. This removes the question of
whether the VmManager instance needs to be configured.

Replace static methods and std::function dispatch with virtual methods.
This removes a level of duplication between the static functions in
VmManager and its subclasses.

Replace singletons with std::unique_ptr return values. This replaces
static state with directly managed memory.

Use a reference instead of a pointer for the CuttlefishConfig input.

Test: launch_cvd
Test: launch_cvd -vm_manager=qemu_cli
Change-Id: Ie27da7133de49329b737e779911c8411a3c6e54c