Add a "client has disconnected" message to device
The client may disconnect while the device is sending forward messages
to it via the signaling server.  Instead of sending an error message
back to the device, which currently causes webRTC to crash, we introduce
a new message type that explicitly tells the device that the client has
disconnected.

There are two places where we need to send the message:
    1. When the connection with the client terminates; and
    2. When the device sends a forward message for the client,

Fixes: 169879650
Bug: 177924439
Test: disconnect webRTC client while device is streaming.
Signed-off-by: Iliyan Malchev <malchev@gmail.com>
Change-Id: I8856bc06a765c126600d29945dbc2c0cf59d0083