Add support for qemu, aarch64 bootloaders
Restructure the makefiles to include a "bootloader.mk" fragment, like
what was done for the kernel modules configuration. Set up aarch64
bootloader support; currently this is only supported on QEMU, so we
just copy the QEMU bootloader twice.

The crosvm bootloader goes to "bootloader" and the QEMU bootloader
goes to "bootloader.qemu". The QEMU bootloader remains unused in this
CL; it will be used in a follow-up change.

Bug: 170433295
Change-Id: I4cd4a34bb351532545b064ec4a5d1fa6bae04e7f