Enable f2fs casefold support with '-data_policy always_create'
When '-data_policy always_create' is passed to launch_cvd, the casefold
feature isn't being enabled on the userdata filesystem.  This is because
in that case the userdata image is created by libcuttlefish_host_config,
and the userdata image produced by the Android build system isn't used.

Fix this by adding the needed option to the mkfs.f2fs command.

Bug: 138322712
Test: 'cat /sys/fs/f2fs/dm-8/features' now shows the casefold feature.
Change-Id: I15991f4e188c8907855c9d3ff58c5f0be989ee7b