cvd_host_package produces *.tar.gz output
In preparation for cvd_host_package to replace host_package.mk, the
output file format of cvd_host_package is changed from zip to tar.gz.
This change is needed because the tar.gz format is expected in a number
of places (e.g. acloud). Changing the output type here is much less
risky than changing the expected file format in all the other places.

Bug: 159685774
Test: m dist out/dist/cvd_host_package.tar.gz
Change-Id: I17d320ba57aa241780b6a258875e7c0e5cf4293a