Install shell_and_utilities_* for first stage console
If the device uses a dedicated recovery partition,
shell_and_utilities_vendor_ramdisk are installed so that sh, toybox
etc. are available under /system/bin in the final ramdisk. They are
installed to the vendor ramdisk instead of the recovery ramdisk so that
they are available in both recovery mode and Android mode (i.e.
regardless of whether the recovery ramdisk is concatenated or not).

If the device installs recovery resources to vendor_ramdisk,
shell_and_utilities_recovery are installed so that sh, toybox
etc. are available under /system/bin in the final ramdisk. First stage
console is started before switching root into /first_stage_ramdisk.
Because recovery resources are moved to vendor ramdisk, sh, toybox etc.
are eventually installed to vendor ramdisk.

This change alone has no effect. By default, TARGET_DEDICATED_RECOVERY
is not set unless it is set from environment variables.
linker.recovery and shell_and_utilities_recovery are installed
by default in base_vendor.mk, so when it is not set, no new modules
are installed.

Bug: 156098440
Test: pass
Change-Id: I453568962b41ebba5c1501fec24a6b3fc2e4d039