Explicitly mark carwatchdog_aidl_interface as missing.
Both android.automotive.watchdog and carwatchdog_aidl_interface
interfaces are not available for phone. Both of these interfaces have
the same package name as they are duplicates and the unittest marks only
one of these interfaces as known missing because the test uses a set of
package names and duplicate entries of the same package name is not
possible to store with a set. Thus add a hack to explictly identify
carwatchdog_aidl_interface and mark it as known missing. This hack will
be removed when carwatchdog_aidl_interface is removed.

Test: The unit test passes for this package.
Bug: b/171422266

Change-Id: Ied33def20c731895eb48b1605e8369b48044c6a9