Don't send frames from vsock hwcomposer when there are no consumers
The vsock hwcomposer is notified by the streamer of whether they want
the frames or not based on how many clients there are connected. The
hwcomposer sends the last frame not previously sent upon receiving the
notification.

Use an unsigned int for the sequential frame number in the
hwcomposer. This number can overflow which is technically undefined
behavior in c++.

Bug: 172587529
Fix: 172684162
Test: locally with vnc and webrtc, check cpu consumption
Change-Id: Id8ac140b540c22e9f996869e94066eb0415be9e3