cvd-host-package is simplified
Build rules for the module type is written using the new PackagingBase.
Traversing the dependencie and copying their output to a staging
directory are now done by PackagingBase. cvd-host-package is only
responsible for the final packaging step of creating tar.gz.

This change has another benefit that the content of the package doesn't
pollute out/soong/host directories.

Bug: 159685774
Bug: 172414391
Test: m dist out/dist/cvd-host_package.tar.gz

Change-Id: I72addea24abf7465bd373dcbaf60ccc5ff47b469