Add Ethernet bridge
Create an Ethernet bridge ("cvd-ebr") alongside the existing wifi bridge
("cvd-wbr") and corresponding tap devices ("cvd-etap-X") for each
instance. This enables a wired ethernet connection to be assigned to
each virtual device and for the networks to be isolated on the host.

If the user wants to bridge with an existing interface, the wifi and
ethernet are bridged together for simplicity, and there is no network
isolation.

Signed-off-by: Alistair Delva <adelva@google.com>