Add Ethernet support to cuttlefish
Add a launcher flag (e.g. "launch_cvd -ethernet=true") which adds
another Ethernet interface to the virtual device. Ethernet devices are
already "buried" for use by wireless, but this new interface is not
wrapped and can be managed by netd. This functionality is useful for
Auto targets.

This mildly refactors some code in allocd to create separate Ethernet
and Wifi bridges and tap devices.

The code is off by default because it requires a new cuttlefish-common
package (0.9.17) and we haven't made the necessary changes to the
networkAttributes to allow Ethernet and Wifi to co-exist, so enabling
Ethernet currently breaks Wifi on phone configurations.

The created ethernet interface will be eth2 for now; once we can enable
this by default, we can move Ethernet back to eth0.

Bug: 172286896
Change-Id: I38ebf259f8eac101d867279f40cc78088a60921d