Reland: Support multi-display in ScreenView and ScreenConnector
Please compare against patchset 1 to see diffs versus the
original aosp/1518579.

The original aosp/1518579 mistakenly disabled the optimization
from aosp/1490174 which stops hwcomposer from sending frames
to the host if there is no vnc/webrtc clients connected. This
reland change adds that back in BroadcastLoop().

Also replaces the present buffer mutex and uint32_t with
atomic<uint32_t>.

Note: this does not enable multi-display yet!

Updates the ScreenConnector interface callback to have a
display number indicating which display the latest frame
is from and updates streaming servers to ignore frames
that are not from the first display.

Adds a helper class to the socket-based ScreenView and
ScreenConnector which encapsulates the existing behavior
of acquiring a buffer for the next frame and marking the
buffer as ready to be sent to host (ScreenView) or to the
streamer (ScreenConnector). Then, create one of these
helpers per display and update the send protocol from
<frame size><frame bytes> to
<display number><frame size><frame bytes>.

Removes the "on frame number after" semantics from the
ScreenConnector interface. The "consume" on ScreenConnector
will handle the case if the streamer registers to receive
a frame after it has already arrived.

Bug: b/173523482
Test: launch_cvd --start_webrtc=true --gpu_mode=guest_swiftshader
Test: launch_cvd --start_vnc_server=true --gpu_mode=guest_swiftshader
Test: launch_cvd --start_webrtc=true --gpu_mode=gfxstream
Test: launch_cvd --start_vnc_server=true --gpu_mode=gfxstream
Test: treehugger avd/avd_boot_test_cloud_tf (x 15 times)
Change-Id: Ic619e6110aeaea298eb20592832d4de342391f28