Strip Vendor Ramdisk of kernel modules
When booting with no bootloader and custom kernels, the vendor boot
ramdisk was previously not preserved. When init was moved to the vendor
ramdisk, cuttlefish no bootloader boot up for custom kernels and kernel
modules was broken.

This change fixes that functionality.

Note that the no-bootloader boot path will be removed as soon as the
bootloader is enabled for all cuttlefish variats (ARM is currently
pending).

Bug: 174952865
Test: local boot with no bootloader (and with custom kernel/ramdisk)
Change-Id: Ia329ca25494d764e63152c5aa713c01a9c79e679