Replace control sockets in the process monitor with a wait() loop.
This removes a file descriptor passed to subprocesses and removes the
chance that subprocesses mess with the control socket by closing it or
writing to it directly.

The process monitor now executes as a forked subprocess rather than a
thread, to avoid waiting on child processes it isn't directed to monitor.
The main downside at the moment is that the subprocess exit callback and
stopper callbacks now execute inside the forked subprocess. That can be
fixed but would make the implementation more complex.

Test: launch_cvd && adb reboot
Test: launch_cvd --daemon && stop_cvd
Bug: 174794655
Change-Id: I1a48545ca3dd4ca33235179d15b9f8b0420d3a1a