Disable Intel HDA driver
The Intel HDA driver probe can claim devices that look like AC-97 with a
DSP, but crosvm doesn't quite implement enough for it to work. The
result is the HDA driver claims the audio device, so the older AC-97
driver can't, which breaks audio.

This appears to be quite a common problem in the real world, so there's
a handy "enable" module parameter which can be set to 0 to disable the
HDA driver probe.

Bug: 174713663
Bug: 175151042
Change-Id: Ie5438661cad6e99b9b95fb39711582939c937c4d