Make Cuttlefish fully compatible with 2021 CTS SIMs.
Add new CTS UICC certificate hash to Cuttlefish's ARF and implement
support for new EAP-SIM and EAP-AKA tests.

It seems EAP-AKA was never implemented before this (but EAP-SIM was),
which just left a wakelock waiting for a RIL response that never came.

In 2021, we'll be releasing an updated CTS UICC specification, and part
of the change will be moving to a new signing certificate. For new
releases, the old signing certificate (aosp-testkey in makefiles) will
no longer be used to grant carrier privileges for
CtsCarrierApiTestCases.

Bug: 178419755
Fix: 179525593
Test: atest CtsCarrierApiTestCases (including testEapSimAuthentication
  and testEapAkaAuthentication)
Test: temporarily switch CtsCarrierApiTestCases to be signed with
  cts-uicc-2021-testkey, ensure suite passes with updated cuttlefish
  modem simulator ARF content

Change-Id: I0df2a774eb3ee9aad2eb62802603088c5bd43910