Add -gdb_port to replace -qemu_gdb
crosvm supports kernel gdb now, but only in TCP/IP mode, so remove the
old qemu_gdb 'flag' and replace it with another flag which is
specifically only to specify the gdb port. This also adds '-S' to the
command line for QEMU, so it waits for the debugger to connect, like
the crosvm support unconditionally does.

On crosvm, the support is still quite limited; only x86_64 is supported,
and only UP (single processor) for now. This change requires the user to
explicitly specify '-cpus 1' so as to not break SMP debugging with QEMU.

Bug: 178531428
Change-Id: Ia07f660c2568f89f256111cb3ab3cdb8640e51f2