[WebRTC UI] Remove the onInitialClick event handler
That even handler was needed because a video elements without controls
is not allowed to play sound in chrome, so the controls were removed
after the first interaction from the user. This is no longer needed
since audio and video now start playing after the user first
interacted with the page (clicks on connect).

Bug: 178735063
Test: locally
Change-Id: I1002d5b8d068e4ae46b36cd9fde6c33b64cd5c67