modem-simulator: keep signal strength percentage
when change different technology (3G, LTE to 5G),
the signal strength should not be reset to 0,
and it seems reasonable to retain the previous
percentage values.

In addition, when updating signal strength,
need to set other signal types to INT_MAX, as
required by guest to indicate invalid values.

For NR, the LTE signal strength should also be
set, as NR relies on LTE as primary.

BUG: 178232863

Test:
check cuttlefish has full signal bar with 5G/3G/2G, in addition to LTE
check goldfish as well

Change-Id: I9be3d8a248d43fd5b0567def1a35797837a9db20