ril: Add missing entry for getHalDeviceCapabilities.
Commit 245f9c4 added a new RIL request
RIL_REQUEST_CONFIG_GET_HAL_DEVICE_CAPABILITIES, but did not add the
corresponding entry to s_configCommands, leading to an out-of-bounds
read in ril.cpp:670 when reading the responseFunction function pointer
(as it happens, the linker laid out a table of function pointers after
s_configCommands, so the problem may not have been noticed without
something like CFI that checks function signatures). Fix the problem
by adding the missing entry to s_configCommands, and add a comment
to try to avoid the problem in the future.

Change-Id: Id611cd532d2453d09a4a26d68fd01654c275a8e4