Use android::base::EndsWith rather than custom suffix testing.
The custom suffix testing had a bug that could cause short filenames to
unexpectedly produce matches for each other. If the string "abc" was
tested if it matched the suffix "zxcv" it would unexpectedly match it.
This is because `expected_pos = file.size() - suffix.size()` would
unintentionally match the result of `file.rfind(suffix)` when
`suffix.size() == file.size() + 1`, because `rfind` returns
std::string::npos on no match, which is also -1.

This problem was previously unlikely with absolute path filenames, but
causes a lot of spurious matches with local file paths (aosp/1582769).
The absolute path filenames were much longer than the local filenames.

Bug: 178677060
Test: Treehugger
Change-Id: I132d3d00dadc21ad8789ca6cb02b313d3b73be88