Avoid checking FLAGS_gpu_mode for true GPU mode
The FLAGS_gpu_mode is "auto" by default, and it gets remapped in the
config object to another setting in various scenarios. This means this
block of code, which had been intended to be optional for swiftshader
or QEMU+swiftshader, was basically being enabled all the time.

It seems like defining the frames_server all the time causes no problem,
but setting the keyboard and touch servers up will break because there
is nothing listening on the other side.

Change-Id: Ia44d6a80000941a5e4a088ca81cf8f8f2398f1e0