Reland "Add enable_rootcanal"
This reverts commit 9c116f275ffd1397015f5c9c15a827554f26926b.

To use BT HAL, the location of root-canal is changed from guest to host.
So if enable-rootcanal is true, run_cvd excute the root-canal binary and
open the port for hci, link and test.

The root-canal does not support arm64 host yet. So it's only applied to
x86_64 host. After supporting root-canl for arm64 host, the root-canal
is enabled for arm64 host

Bug: 178702846
Bug: 181203470
Test: lunch aosp_cf_arm64_phone-userdebug && m -j
Test: lunch aosp_cf_x86_64_phone-userdebug && m -j
Test: launch_cvd --enable-rootcanal=true

Change-Id: I8149e3c1330b95ae443e8f18d738eaebf1bddfc5