Move vhost-net option up
This ignored a comment that explained that --bios needed to be the last
parameter (which is actually because if --bios is not specified, the
kernel image will be instead, which should be last).

Move it to the top of the block to prevent it falling off again in the
future, since this is more of a "global" option than
device-block-specific.

Change-Id: I5033b56e504e8423cb53985d78d76f0541b74bb1