Swap FRP and SD-Card and make SD-Card optional
Currently the FRP is just 1MB and not a priority to disable, but the
SD-Card can be removed even from physical devices, so there is little
reason to require this device to always be present. Reshuffle the FRP
and SD-Card for now, so /dev/vdb will always be there (and FRP) and
/dev/vdc may optionally not be there.

This adds a new '-use_sdcard=false' to disable the SD-Card feature.

Change-Id: Iaaa653c2b13445f614a1a2fda3eee616a22d9847