Introduce --protected_vm, drop some devices
When specified, pass --protected-vm to crosvm.

Specifying this flag implicitly remove some block devices and partitions
that do not make sense in a secure VM. This change drops frp,
access-kregistry and pstore. The only boot breakage comes from dropped
frp, which is checked by the framework. Note that using --protected-vm
is NOT SUPPORTED with the normal cuttlefish builds.

Change-Id: I3b494d39ababcf1a6eb2f378baa9134795249b46