socket_vsock_proxy is built for microdroid
This change introduces socket_vsock_proxy_microdroid, which is
identical to the existing socket_vsock_proxy except for the `vendor:
true` bit. The former is put to the `com.android.virt` APEX to relay the
adb connection between adb client and the adbd daemon running in
microdroid. We can't use the existing one in the APEX due to `vendor:
true` bit. The APEX is not built for vendor.

Bug: 180896120
Bug: 181728474
Test: m com.android.virt
Change-Id: Ib65c2995377d8e1d017918b805bc552240e338c6