Don't create uboot_env.img for the on-device use cases.
uboot_env.img has been created by assemble_cvd. It has information about
boot mode and kernel commandline parameters. However, in case of pKVM
use case, creating the partition on the device is considered insecure
because the partition is not in the verified boot chain; one can create
an arbitrary uboot_env.img with harmful kernel cmdline parameters and
still be able to boot the VM.

With this change, assemble_cvd doesn't attempt to create the partition
when it's built for Android.

Bug: 180571439
Bug: 181812679
Test: try booting microdroid following the instruction in
package/modules/Virtualization/microdroid/README.md

Change-Id: I96244108b0cebb75f62a8a4dc1c88969dcbd5f6b