Fix metrics prompt causing 100% cpu usage
The issue was caused by cin.get, which immediately
returns when there's no TTY present and doesn't alter
the provided char.  This caused an infinite loop.

The fix is to check the eof flag on the cin stream,
which is set when there's no TTY present, and default
to 'n'.

The downside to this is there will not be any metrics
collection when launch_cvd is run with no TTY, unless
-report_anonymous_usage_stats='y' is specified as an
argument to launch_cvd.

Bug: 181600535
Test: local build
Change-Id: Ic7eafcb7337ef56d75cafcb76f545b4af27224da