Don't repack on Android
Currently, assemble_cvd repacks boot and/or vendor_boot images. It does
that either when the kernel or ramdisk is given (so that the boot images
are composed of the given kernel and ramdisk), or when simply
bootloader option is on - just to convert vendor_boot to the V4 format.

However, this behavior can't be supported on Android. The repack process
depends on host tools like unpack_bootimg, mkbootimg, etc. The tools are
currently not built for Android and some (e.g. unpack_bootimg) can't as
they are python scripts.

This change disables the repacking feature for the Android build of
assemble_cvd. When repacking is requested (by providing a kernel or
ramdisk with the bootloader option), it rejects the request.

Bug: 181843313
Test: run micrdroid following
packages/module/Virtualization/microdroid/README.md

Change-Id: I3c4e86b7cc73444e99f26ed304d33d32f6585c35