Change androidboot.hardware parameter to hardware
Because bootconfig does not allow keys to contain subkeys and values,
androidboot.hardware cannot coexist androidboot.hardware.* parameters.

"androidboot.hardware" is being changed to "hardware" when present
in bootconfig.

If the kernel does not support bootconfig, that parameter needs to be
changed back to "androidboot.hardware". assemble_cvd is responsible for
this replacement.

Test: launch_cvd with 4.19 and 5.10 kernels
Bug: 173815685

Change-Id: Ib214826b3fb1103f2f4554021a00df59445f7064