Add grub-efi dependency to cuttlefish-common
Ensure that at least the bootloader binaries for the native architecture
are installed, so that Other OS mode can work. For cases where
non-native targets are supported (such as QEMU TCG) the user will have
to install the extra needed bootloader binaries manually.

This inverts some checks for '!amd64' to be 'arm64', though in practice
it shouldn't matter too much as cuttlefish cannot run on 32-bit arm or
32-bit x86 presently.