Fix support for QEMU on non-native architectures.
Add the target architecture to the cuttlefish config, inferred
from the kernel config. Fix several assumptions about the host
architecture being the same as the target. When running on a non-native
architecture, default to QEMU, emulate all available CPU features
and increase the timeout multiplier to compensate for the slower
execution speed.

Bug: 161538883
Change-Id: I57ffb1f2af3bdf7ede4db96a27455fef93c14b99