Support auto-generation of ESP, add otheros_root
The existing code allowed a user to inject an ESP, but building a good
one is quite onerous, as you need to find the right GRUB binaries, put
them in the right place, make a grub.cfg to chain load, and so on.
It's better in most cases if assemble_cvd takes care of this process.

This change uses the new mtools binary to modify a FAT32 partition
without requiring root, and pulls GRUB from the host machine. A suitable
grub.cfg is injected into the ESP as well. GRUB is taken from the host
because it is difficult to build, and we can't add it to AOSP right now.

This new partition can boot a kernel / initramfs.img directly from a
static location in the ESP ("/vmlinuz" and "/initrd.img"), or it can
search for another grub installation and chain-load it.

This change removes the old -use_esp flag, and an ESP will be generated
when needed by the otheros flags. Three new flags are added:

  -otheros_kernel_path     The path to a raw vmlinuz
  -otheros_initramfs_path  The path to an initramfs.img
  -otheros_root_image      The path to a root filesystem image.

If the root filesystem is not specified, the other options have no
effect.

Bug: 175204891
Change-Id: I2c4c4316dc1e4e7bae08893e4feff4c4fd9baf9c