Remove std::vector from ConfigureBootDevices
The boot devices must all be provided at the same time with the same
boot property, and most of the time we use a string builder to hexify
the PCI IDs, so we might as well build the comma separated list there
at the same time. Remove the never-used vector of strings.

Bug: 183006588
Change-Id: I7ae4033d0a400d33b6592b3aee907896c33b0fc0