Adds state:posture mapping to aosp_cf_x86_64_foldable.
Includes a new HALF_OPENED state with no conditions at a high
id value so that it does not take precedence over the other states.
This state will use the hinge sensor when that is available.

Also sets the config_display_features for this AOSP target so that
apps know the location of the hinge.

Bug: 183447987
Test: launch_cvd for aosp_cf_x86_64_foldable
      Launch an app that uses postures. Use `cmd device_state state 3`
      to force the HALF_OPENED state. Observe app changes.
Change-Id: Ibb7739f38ce92342a2d108228ccc963ca2cb3178