assemble_cvd: add read_only to ImagePartition
ImagePartition struct is mapped to ComponentDisk. While ComponentDisk
has read_write_capability, ImagePartition doesn't and
CompositeDiskBuilder uses "READ_WRITE" as hard-coded.

This change adds read_only member to ImagePartition so that when it is
set to true, ComponentDisk's read_write_capability is set accordingly.

For now, it doesn't change the original behavior because the added field
is set to "false" by default and it is mapped to "READ_WRITE".

This is a preparation step towards extracting the feature
CompositeDiskBuilder so that it can be used by others with that
capability.

Bug: n/a
Test: launch_cvd works okay (which calls assemble_cvd internally)
Change-Id: I4abb7c7b5435fc034227e9bfd74e3d6c5bfa354a