assemble_cvd: composite disks with absolute paths
A composite disk is a this wrapper to other images. However, there's no
explicit rules regarding the references.

This change makes it explicit by putting absolute paths in a composite
disk.

Alternatively, we could make crosvm support "relative paths" in
composite disks, which would be a separate issue. And putting absolute
paths in a composite disk should do no harm regardless of whether crosvm
supports rel-paths or not.

Bug: n/a
Test: launch_cvd
Change-Id: Ib45ea171a4f7c550b2c83bc42b02af9458f7a210