Concat all vendor ramdisks when repacking vendor_boot
vendor_boot v4 supports packaging multiple "fragmented" vendor ramdisks.
|unpack_bootimg| would unpack these individual ramdisks as
"vendor_ramdisk0", "vendor_ramdisk1" ...

To fully support vendor_boot v4, we would have to unpack-strip-repack
each of these individual ramdisks and then reconstruct the entire
mkbootimg command. This would be a lot of work with little benefit, as
the current bootflow of CF bootloader doesn't treat fragmented vendor
ramdisks differently. It just loads all of the ramdisk fragments as a
whole as the initramfs.

So here I just concat all of these "vendor_ramdisk*" files into a single
chunk of ramdisk, and repack the vendor_boot with the single ramdisk,
which is the concatenation of all vendor ramdisk fragments.

Apply these changes to the BoardConfig.mk to create vendor_boot with
multiple ramdisks, where the vendor_ramdisk DLKMs are packaged in a
standalone ramdisk apart from the platform and recovery ramdisk:

  ```
  +BOARD_VENDOR_RAMDISK_FRAGMENTS := dlkm
  +BOARD_VENDOR_RAMDISK_FRAGMENT.dlkm.KERNEL_MODULE_DIRS := top
  ```

Bug: 184260498
Test: Create a vendor_boot with multiple ramdisks, CF still boots.
Test: Forrest run boot with 4.19 and mainline kernels.
Change-Id: Id7939592aceff57f34747bfea5f95899f3afc3c6