Move kernel module options from bootconfig to modules.options
The bootconfig elements are stored in the vendor boot partition.
By making this change - the builder will shift them from the
bootconfig in the vendor boot partition to the ramdisk in the
same partition. Meaning that any version of cuttlefish launched
with a launcher from this point forward will still have these
settings present in their vendor boot image. i.e. this change is
backwards compatible.

Also added a comment clarifying the blocklist location.

Bug: 182417593
Test: Boot CF and check logcat; verify these logs do show up:
  init    : Loading module /lib/modules/mac80211_hwsim.ko with args
  'radios=0'
  init    : Loading module /lib/modules/snd-hda-intel.ko with args
  'enable=0'
Change-Id: I6bb31a0c8591cbe60275e8ce256ee2eb0829e686