libimage_aggregator: fill gaps in composite disk image
When creating a composite image with a file which is not aligned to 4K
boundary (say, 100bytes) its resulting partition will have a gap (or a
hole), which causes an error when reading the block.

It's fixed by putting the fillter disk (reusing the gpt header image) in
those gaps.

Note that these filler disks are added only for the "read-only" disks.
An unaligned file with "read-write" capability is an error.

Bug: 184605708
Test: mk_cdisk cfg.json out.img
  with a single partition with a file of 100byte.
  pass the out.img to crosvm and try to read it.
Change-Id: I86962d065fb6a8d8f71e9b33c2c38989867fc390