Fix + install guest/libs/wpa_supplicant_8_lib as lib_driver_cmd_simulated_cf
Brings this code somewhat up to date. Reintroduces Android.mk files to
make sure this code builds.

This library is necessary to support "vendor commands" like finding the
MAC address.

Due to difficulties with reusing goldfish's implementation of the same
name, this is renamed to lib_driver_cmd_simulated_cf to avoid conflicts.

Test: Enable wifi + atest CtsDevicePolicyManagerTestCases:MixedDeviceOwnerTest#testWifi
Test: build/soong/build_test.bash --dist
Bug: 137088260
Change-Id: I26394c6810eedb4a1156f1844843824e71b14680