Change recovery serial console to a root shell
It is already running as the ROOT uid, however it is running in the
"u:r:shell:s0" secontext, which renders it as useful as a regular shell.

Change the secontext to "u:r:su:s0" so that it behaves more like a root
shell like "adb root && adb shell".

Test: $ launch_cvd -daemon -start_webrtc -console
  $ adb reboot recovery
  $ minicom -p $(realpath ~/cuttlefish_runtime.1/console)
    > id
    uid=0(root) gid=0(root)
    groups=0(root),1007(log),2000(shell),3009(readproc)
    context=u:r:su:s0
Change-Id: Ibb3a3dde85a1b40bce512bd41e92064c76d3a846