Update ScreenConnector interface to set callback once
... instead of on a per frame basis. With multi-display,
ScreenConnector would frequently miss and drop frames from not
setting the callback in time.

Note that this slightly changes the behavior of Crosvm and
the VNC streaming server (*not WebRTC*) when there is no client
connection.

- Prior to this change, Crosvm's Wayland display would continuously
send frames to Cuttlefish's Wayland server (which runs inside of
the VNC streaming server). When no clients are connected,
the ScreenConnectorQueue becomes full after 2 frames and the
AndroidFrameFetchingLoop() thread becomes blocked waiting for the
ScreenConnectorQueue to not be full. The AndroidFrameFetchingLoop()
would not re-configure the OnNextFrame() callback and the excess
frames would be dropped in Cuttlefish's Wayland server.

- After this change, Crosvm's Wayland display will not continuously
send frames to Cuttlefish's Wayland server (which runs inside of
the VNC streaming server). When no clients are connected,
the ScreenConnectorQueue becomes full after 2 frames and the
*Wayland server's thread* becomes blocked waiting for the
ScreenConnectorQueue to not be full while trying to run the
process frame callback. This is still okay as Crosvm's
Wayland display sees that Cuttlefish's Wayland server has not
finished using any buffers and drops the next frame.

This doesn't change what VNC clients see but wanted to call it
out. Also note that VNC clients see very old frames when a user
finally connects both before and after this change - this is
problematic as a VNC client will see a boot animation frame if
they connect to VNC after the device has reached a steady state
and isn't producing new frames.

This doesn't change behavior when running with WebRTC as the WebRTC
server continues to process frames even when a client is not
connected.

Bug: b/186633303
Test: launch Cuttlefish and start webrtc
Test: launch Cuttlefish and start vnc
Change-Id: I03e5f11d8647f1f035b8dcb03481751a82826d50