Set Cuttlefish KeyMint message version
The remote KeyMint implementation in cuttlefish relies on an underlying
Keymaster component for most work. However, the RemoteKeymaster had
hard-coded its message version to a pre-KeyMint value. Update
RemoteKeymaster so that the version can be set explicitly.

Update the Cuttlefish guest and host KeyMint components to exlpicitly
use the KEYMINT_1 message version.

Test: VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I327d48f70e593964202298ac5d95673d00a9c10e