Fix TPM key blob tagging
The SplitEnforcedProperties function was a bit too overzealous in what
it allowed to be considered as software key authorization tags. Limit
what is allowed to be used as a software key tag.

Additionally, the tag filtering was done in two places -- this unifies
it all into one function. There is no default case, so any new tags
that are added in the future must be explicitly accounted for.

Bug: 182928606
Test: VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I93df9caf771199962f4e81643fff0fbe5fb5ad26