Always use the `/url` endpoint to download files.
The alternative path was saving an error message and considering it a
success when there was an error from the server side. This approach more
consistently reports an error back to its caller when there is a
server-side issue.

There is other code that depends on this error-handling working
correctly, specifically the decision to download the "bzImage" or
"Image" files for a kernel build.

Bug: 169261091
Bug: 191309622
Bug: 191302402
Test: fetch_cvd with an internal build
Test: fetch_cvd with and without credentials for an aarch64 kernel build
Change-Id: I1bf466984bfe0e603a961113563dc8ed56efff38