Report that attestation is supported.
Based on information from eranm@ and zeuthen@ this file is required to
be present when the device supports attestation, and not present when
attestation is supported.

As cuttlefish here uses the default keymint implementation which does
support attestation, copy this file into the vendor partition.

Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.MixedProfileOwnerTest#testKeyManagement -- --abi x86_64
Bug: 177800966
Merged-In: Ie1553a0cc4f387f588dbb0c1bf889ad888365f0f
Change-Id: Ie1553a0cc4f387f588dbb0c1bf889ad888365f0f