Add a health host binary to send battery parameters to crosvm
This only sends the battery parameters that crosvm currently
supports, namely: status, health, present, capacity, ac online.
This will need to be expanded once crosvm supports more.

The sysfs paths can be read directly after using this binary,
but the intent is to read those from the health HAL so the
framework can use it.

Bug: 181341330
Test: local build and boot
Change-Id: Ice57ab793b9b3ce7acfa54e70824cae573923a65