Fix multi-SIM device instantiation
The first issue was that the tokenizing of the ports
from the cuttlefish config had an off by one error,
which caused a truncation of the port value. That
port was already in use, which caused an error on
the first modem simulator instantiation.

The second issue after this was resolved was that
the cuttlefish config was built with the exact same
port number assigned to each modem simulator.
Therefore, this failed when the second modem
simulator was instantiated.

Bug: 169337277
Test: local build and boot
Change-Id: Ic27e02de07eaa17a28645304d29ec84336179047