Fix CF support for application-bound keys
Add support for additional authenticated data to HmacSerializable so
that integrity checks may include data that is not necessarily
encrypted/included in the target payload.

Include application data/id in the AAD on key MACs so that they are
treated as secrets, and must be passed in so that a key may be
unwrapped and used.

Test: VtsAidlKeyMintTargetTest
Bug: 182928606
Change-Id: I9621989d8dbf24930a881b8f47161e4f1b12990f