Change userdata filesystem type if different than current
If the user already has a userdata image that is of f2fs
type and then specifies f2fs as the userdata_format, the
userdata image does not get recreated. However, if the
formats between what the userdata image is and what the
user requests (per `userdata_format`) differ, then the
userdata image is formatted.

Bug: 142424832
Test: launch_cvd --userdata_format=f2fs
Test: launch_cvd --userdata_format=ext4
Change-Id: I659d21cb4f6b6caa5644a6269c22bbf3eeb0d06c